refactor nav and preview components
