+++ src/components/generalComponents/NavLinks.Map.css
@@ -26,14 +26,14 @@
   bottom: 0;
   background-color: rgba(10, 10, 10, 0.9);
   box-shadow: 0 0 5px 0 #111111;
-  padding: 40px;
+  padding-top: 5rem;
   z-index: 1;
   transform: translateX(100%);
   transition: transform 0.5s ease-out;
 }
 .map-nav-links li {
   opacity: 0;
-  margin: 20px 0;
+  margin: 2.5rem;
   width: fit-content;
 }
 .map-nav-links a {
@@ -68,8 +68,5 @@
   .map-burger div {
     width: 20px;
     margin: 2px;
-  }
-  .map-nav-links {
-    font-size: 90%;
   }
 }
