+++ src/components/generalComponents/NavLinks.css
@@ -1,0 +1,78 @@
+.nav-links {
+  margin: 0;
+  padding: 0;
+  color: rgba(255, 255, 255, 0.9);
+}
+.nav-links li {
+  list-style: none;
+  width: fit-content;
+  display: inline-block;
+  margin-left: 25px;
+}
+.nav-links a {
+  text-decoration: none;
+  letter-spacing: 1px;
+}
+.toggle .line1 {
+  transform: rotate(-45deg) translate(-5px, 6px);
+}
+.toggle .line2 {
+  opacity: 0;
+}
+.toggle .line3 {
+  transform: rotate(45deg) translate(-5px, -6px);
+}
+.burger {
+  display: none;
+}
+.burger div {
+  width: 25px;
+  height: 3px;
+  background-color: #ededed;
+  margin: 5px;
+  transition: all 0.3s ease;
+}
+.underline-hover::after {
+  content: "";
+  display: block;
+  width: 0;
+  height: 2px;
+  transition: width 0.3s;
+  background-color: #ededed;
+}
+.underline-hover:hover::after {
+  width: 100%;
+}
+
+@keyframes navLinkFade {
+  from {
+    opacity: 0;
+    transform: translateX(50px);
+  }
+  to {
+    opacity: 1;
+    transform: translateX(0);
+  }
+}
+
+@media (max-width: 870px) {
+  .nav-links {
+    position: absolute;
+    right: 0;
+    top: 5rem;
+    height: calc(100vh - 5rem);
+    background-color: rgba(10, 10, 10, 0.9);
+    width: 50%;
+    transform: translateX(100%);
+    transition: transform 0.5s ease-out;
+  }
+  .nav-links li {
+    opacity: 0;
+    margin: 2.5rem;
+    display: block;
+  }
+  .burger {
+    display: block;
+    cursor: pointer;
+  }
+}
