+++ src/components/generalComponents/StoryPreview.Profile.css
@@ -1,4 +1,4 @@
-.story-wrapper {
+.profile-story-wrapper {
   width: 100%;
   display: grid;
   grid-template-columns: 30% 70%;
@@ -7,10 +7,10 @@
   border: 1px solid rgba(255, 255, 255, 0.3);
   border-radius: 2px;
 }
-.story-wrapper:not(:last-child) {
+.profile-story-wrapper:not(:last-child) {
   margin-bottom: 1rem;
 }
-.story-wrapper .story-image {
+.profile-story-wrapper .story-image {
   height: 200px;
   width: 100%;
   display: flex;
@@ -19,20 +19,20 @@
   overflow: hidden;
   border-radius: 2px 0 0 2px;
 }
-.story-wrapper .story-image img {
+.profile-story-wrapper .story-image img {
   height: 100%;
 }
-.story-wrapper .story-detail {
+.profile-story-wrapper .story-detail {
   padding: 8px 1rem;
 }
-.story-wrapper .story-detail p:not(:last-child) {
+.profile-story-wrapper .story-detail p:not(:last-child) {
   margin-bottom: 1rem;
 }
-.story-wrapper .story-detail p:not(.story-title) {
+.profile-story-wrapper .story-detail p:not(.story-title) {
   font-size: 14px;
   font-weight: 300;
 }
-.story-wrapper .story-detail .story-hyperlink {
+.profile-story-wrapper .story-detail .story-hyperlink {
   width: 100%;
   text-overflow: ellipsis;
   white-space: nowrap;
@@ -40,14 +40,14 @@
 }
 
 @media (max-width: 480px) {
-  .story-wrapper {
+  .profile-story-wrapper {
     display: block;
   }
-  .story-wrapper .story-image {
+  .profile-story-wrapper .story-image {
     border-radius: 2px 2px 0 0;
     height: 130px;
   }
-  .story-wrapper .story-image img {
+  .profile-story-wrapper .story-image img {
     width: 100%;
     height: auto;
   }
